requirements.txt
annotated-types==0.7.0
anyio==4.9.0
awsebcli==3.23.1
beautifulsoup4==4.13.3
blessed==1.20.0
blinker==1.9.0
boto3==1.37.23
botocore==1.37.23
bs4==0.0.2
cachetools==5.5.2
cement==2.10.14
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coverage==7.8.0
deep-translator==1.11.4
distro==1.9.0
et_xmlfile==2.0.0
Flask==3.1.0
google-api-core==2.24.2
google-auth==2.39.0
google-cloud-core==2.4.3
google-cloud-translate==3.20.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
lxml==5.3.1
MarkupSafe==3.0.2
mwparserfromhell==0.6.6
openai==1.72.0
openpyxl==3.1.5
packaging==24.2
pathspec==0.12.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rsa==4.9.1
s3transfer==0.11.4
semantic-version==2.10.0
setuptools==78.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
termcolor==2.5.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.0
urllib3==1.26.20
wcwidth==0.2.13
Werkzeug==3.1.3
